+++ Sources/KSCrash/llvm/Support/Raygun_Casting.h
