+++ Sources/KSCrash/swift/Basic/Raygun_LLVM.h
@@ -22,7 +22,7 @@
 // dependencies.
 // Casting.h has complex templates that cannot be easily forward declared.
 //#include "llvm/Support/Casting.h"
-#include "Casting.h"
+#include "Raygun_Casting.h"
 // None.h includes an enumerator that is desired & cannot be forward declared
 // without a definition of NoneType.
 //#include "llvm/ADT/None.h"
